% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BNPmix_VI_LB}
\alias{BNPmix_VI_LB}
\title{C++ function - compute the VI lower bound}
\arguments{
\item{M}{a matrix (r x n), r number of replications, n number of observations}

\item{psm_mat}{a posterior similarity matrix}
}
\description{
C++ function - compute the VI lower bound
}
\examples{
{
  M <- matrix(c(1,1,1,2,1,1,2,2,1,1,2,1,1,1,1,2), ncol = 4)
  psmM <- BNPmix_psm(M)
  BNPmix_VI_LB(M, psmM)
}

}
\keyword{internal}
