% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cICS_mv_L}
\alias{cICS_mv_L}
\title{C++ function to estimate Pitman-Yor multivariate mixtures via importance conditional sampler}
\arguments{
\item{data}{a matrix of observations}

\item{grid}{matrix of points to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{S20}{variance of location component}

\item{S0}{parameter of scale component}

\item{n0}{parameter of scale component}

\item{m1}{mean of hyperdistribtion of m0}

\item{k1}{scale factor of hyperdistribtion of m0}

\item{theta1}{df of hyperdistribtion of S20}

\item{Theta1}{matrix of hyperdistribution of S20}

\item{strength}{strength parameter}

\item{napprox}{number of approximating values}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{discount}{second parameter of PY}

\item{print_message}{print the status}

\item{light_dens}{if TRUE return only the posterior mean of the density}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}
}
\description{
C++ function to estimate Pitman-Yor multivariate mixtures via importance conditional sampler
}
\keyword{internal}
