% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clean_partition}
\alias{clean_partition}
\title{C++ function - clean the partition matrix}
\arguments{
\item{M}{a matrix (r x n), r number of replications, n number of observations}
}
\description{
C++ function - clean the partition matrix
}
\examples{
{
  M <- matrix(c(1,1,1,3,1,1,4,4,1,1,3,1,1,3,1,1), ncol = 4)
  clean_partition(M)
}

}
\keyword{internal}
