% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSLI_L}
\alias{cSLI_L}
\title{C++ function to estimate Pitman-Yor univariate mixtures via slice sampler - LOCATION}
\arguments{
\item{data}{a vector of observations}

\item{grid}{vector to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{s20}{variance of location component}

\item{a0}{parameter of scale component}

\item{b0}{parameter of scale component}

\item{m1}{hyperparameter, mean of distribution of m0}

\item{k1}{hyperparameter, scale factor of distribution of m0}

\item{a1}{hyperparameter, shape of distribution of s20}

\item{b1}{hyperparameter, rate of distribution of s20}

\item{mass}{mass parameter}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{sigma_PY}{second parameter of PY}

\item{print_message}{print the status}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}
}
\description{
C++ function to estimate Pitman-Yor univariate mixtures via slice sampler - LOCATION
}
\keyword{internal}
