% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSLI_mv}
\alias{cSLI_mv}
\title{C++ function to estimate Pitman-Yor multivariate mixtures via slice sampler - LOCATION SCALE}
\arguments{
\item{data}{a matrix of observations}

\item{grid}{matrix of points to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{k0}{tuning parameter of variance of location component}

\item{S0}{parameter of scale component}

\item{n0}{parameter of scale component}

\item{m1}{mean of hyperprior distribution of m0}

\item{S1}{covariance of hyperprior distribution of m0}

\item{tau1}{shape parameter of hyperprior distribution of k0}

\item{tau2}{rate parameter of hyperprior distribution of k0}

\item{theta1}{df of hyperprior distribution of S0}

\item{Theta1}{matrix of hyperprior distribution of S0}

\item{mass}{mass parameter}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{sigma_PY}{second parameter of PY}

\item{print_message}{print the status}

\item{light_dens}{if TRUE return only the posterior mean of the density}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}
}
\description{
C++ function to estimate Pitman-Yor multivariate mixtures via slice sampler - LOCATION SCALE
}
\keyword{internal}
