% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSLI_mv_MKR_L}
\alias{cSLI_mv_MKR_L}
\title{C++ function to estimate Pitman-Yor multivariate mixtures via slice sampler - PRODUCT KERNEL}
\arguments{
\item{data}{a matrix of observations}

\item{grid}{matrix of points to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{k0}{vector, scale parameters for the location component}

\item{a0}{vector, parameters of scale component}

\item{b0}{vector, parameters of scale component}

\item{m1}{means of hyperdistribution of m0}

\item{s21}{variances of hyperdistribution of m0}

\item{a1}{shape parameters of hyperdistribution of b0}

\item{b1}{rate parameters of hyperdistribution of b0}

\item{strength}{strength parameter}

\item{napprox}{number of approximating values}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{discount}{second parameter of PY}

\item{print_message}{print the status}

\item{light_dens}{if TRUE return only the posterior mean of the density}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}
}
\description{
C++ function to estimate Pitman-Yor multivariate mixtures via slice sampler - PRODUCT KERNEL
}
\keyword{internal}
