% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPmixMulti.R
\name{DPmixMulti}
\alias{DPmixMulti}
\title{Estimating the posterion mean of a Dirichlet Process Mixture Model via marginal Gibbs sampler.
The base measure is specified with independent component.}
\usage{
DPmixMulti(nsim = NULL, nburn = NULL, napprox = 100, data = NULL,
  grid = NULL, conf_start = NULL, mu_start = NULL, Lambda_start = NULL,
  theta = NULL, m0 = NULL, B0 = NULL, nu0 = NULL, sigma = NULL,
  b1 = NULL, B1 = NULL, m1 = NULL, M1 = NULL, s1 = NULL, S1 = NULL,
  t1 = NULL, t2 = NULL, k1 = NULL, nupd = NULL, fix = FALSE,
  dep = FALSE, seed = 42)
}
\arguments{
\item{data}{A multivariate dataset.}

\item{grid}{An expanded grid of points to evaluating the density.}

\item{MCMC_param}{A list of MCMC parameters}
}
\value{
A list of
}
\description{
Estimating the posterion mean of a Dirichlet Process Mixture Model via marginal Gibbs sampler.
The base measure is specified with independent component.
}
