% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAR_mv_P}
\alias{MAR_mv_P}
\title{C++ function to estimate Pitman-Yor multivariate mixtures via marginal sampler - LOCATION SCALE}
\arguments{
\item{data}{a matrix of observations}

\item{grid}{matrix of points to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{k0}{vector, scale parameters for the location component}

\item{a0}{vector, parameters of scale component}

\item{b0}{vector, parameters of scale component}

\item{m1}{means of hyperdistribution of m0}

\item{s21}{variances of hyperdistribution of m0}

\item{tau1}{shape parameters of hyperdistribution of k0}

\item{tau2}{rate parameters of hyperdistribution of k0}

\item{a1}{shape parameters of hyperdistribution of b0}

\item{b1}{rate parameters of hyperdistribution of b0}

\item{mass}{mass parameter}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{sigma_PY}{second parameter of PY}

\item{print_message}{print the status}

\item{light_dens}{if TRUE return only the posterior mean of the density}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}
}
\description{
C++ function to estimate Pitman-Yor multivariate mixtures via marginal sampler - LOCATION SCALE
}
\keyword{internal}
