% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMCMC.R
\docType{methods}
\name{plot,modCond-method}
\alias{plot,modCond-method}
\title{modCond object plot}
\usage{
\S4method{plot}{modCond}(x, col = "#0037c4")
}
\arguments{
\item{x}{object of class \code{modCond}.}

\item{col}{the color of the density.}
}
\description{
\code{plot} method for class \code{modCond}.
}
