% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condDDP.R
\docType{methods}
\name{plot,modCondDep-method}
\alias{plot,modCondDep-method}
\title{modCondDep object plot}
\usage{
\S4method{plot}{modCondDep}(x, ncol = 1)
}
\arguments{
\item{x}{object of class \code{modCondDep}.}

\item{ncol}{number of column in the final plot.}
}
\description{
\code{plot} method for class \code{modCondDep}
}
