% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condMCMCmv.R
\docType{methods}
\name{plot,modCondMv-method}
\alias{plot,modCondMv-method}
\title{modCondMv object plot}
\usage{
\S4method{plot}{modCondMv}(x, dimension = c(1, 2), col = "#0037c4")
}
\arguments{
\item{x}{object of class \code{modCondMv}.}

\item{dimension}{a vector of two elements. If the elements take the same value, a plot of the
marginal density is provided, otherwise a bidimensional plot of the corresponding dimensions.}

\item{col}{the color of the density.}
}
\description{
\code{plot} method for class \code{modCondMv}
}
