% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{print.BNPdens}
\alias{print.BNPdens}
\title{BNPdens print method}
\usage{
\method{print}{BNPdens}(x, ...)
}
\arguments{
\item{x}{an object of class \code{BNPdens};}

\item{...}{additional arguments.}
}
\description{
The \code{BNPdens} method prints the type of a \code{BNPdens} object.
}
\examples{
data_toy <- c(rnorm(100, -3, 1), rnorm(100, 3, 1))
grid <- seq(-7, 7, length.out = 50)
est_model <- PYdensity(y = data_toy, mcmc = list(niter = 100,
                      nburn = 10, napprox = 10), output = list(grid = grid))
class(est_model)
print(est_model)
}
