% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PYcalibrate}
\alias{PYcalibrate}
\title{Pitman-Yor prior elicitation}
\usage{
PYcalibrate(Ek, n, discount = 0)
}
\arguments{
\item{Ek}{prior expected number of cluster;}

\item{n}{sample size;}

\item{discount}{discount parameter; default is set equal to 0, corresponding to a Dirichlet process prior.}
}
\value{
A named list containingtthe values of strength and discount parameters.
}
\description{
The function \code{PYcalibrate} elicits the strength parameter of the Pitman-Yor
process, given the discount parameter and the prior expected number of clusters.
}
\examples{
PYcalibrate(5, 100)

PYcalibrate(5, 100, 0.5)

}
