% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{summary.modCond}
\alias{summary.modCond}
\title{modCond summary method}
\usage{
\method{summary}{modCond}(object, ...)
}
\arguments{
\item{object}{an object of class modCond}

\item{...}{additional arguments to be passed}
}
\description{
modCond summary method
}
\examples{
data_toy <- c(rnorm(100, -3, 1), rnorm(100, 3, 1))
grid <- seq(-7, 7, length.out = 50)
est_model <- condMCMC(data = data_toy, grid = grid, niter = 1000,
                      nburn = 100, napprox = 100, nupd = 100)
class(est_model)
summary(est_model)

}
