% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPmix-package.R
\docType{package}
\name{BNPmix-package}
\alias{BNPmix}
\alias{BNPmix-package}
\title{BNPmix: Algorithms for Pitman-Yor Process Mixtures}
\description{
Contains different algorithms to both univariate and multivariate Pitman-Yor process mixture models, and Griffiths-Milne Dependent Dirichlet process mixture models. Pitman-Yor process mixture models are flexible Bayesian nonparametric models to deal with density estimation. Estimation could be done via importance conditional sampler, or via slice sampler, as done by Walker (2007) <doi:10.1080/03610910601096262>, or using a marginal sampler, as in Escobar and West (1995) <doi:10.2307/2291069> and extensions. The package contains also the procedures to estimate via importance conditional sampler a GM-Dependent Dirichlet process mixture model.
}
\keyword{internal}
