% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPmix-package.R
\docType{package}
\name{BNPmix-package}
\alias{BNPmix}
\alias{BNPmix-package}
\title{BNPmix: Algorithms for Pitman-Yor process mixtures}
\description{
A package containing different algorithms to both univariate and multivariate Pitman-Yor mixture models. Estimation could be done via importance conditional sampler, slice sampler or marginal sampler. The package contains also the procedures to estimate via importance conditional sampler a Griffiths-Milne Dependent Dirichlet process mixture model.
}
\keyword{internal}
