% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPmix-package.R
\docType{package}
\name{BNPmix-package}
\alias{BNPmix}
\alias{BNPmix-package}
\title{BNPmix: Bayesian Nonparametric Mixture Models}
\description{
Functions to perform Bayesian nonparametric univariate and multivariate density estimation and clustering, by means of Pitman-Yor mixtures, and dependent Dirichlet process mixtures for partially exchangeable data. See Corradin et al. (2021) \doi{10.18637/jss.v100.i15} for more details.
}
\author{
\strong{Maintainer}: Riccardo Corradin \email{riccardo.corradin@gmail.com}

Other contributors:
\itemize{
  \item Antonio Canale [contributor]
  \item Bernardo Nipoti [contributor]
}

}
\keyword{internal}
