% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNPmix-package.R
\docType{package}
\name{BNPmix-package}
\alias{BNPmix}
\alias{BNPmix-package}
\title{BNPmix: Bayesian Nonparametric Mixture Models}
\description{
Functions to perform Bayesian nonparametric univariate and multivariate density estimation and clustering, by means of Pitman-Yor mixtures, and Dependent Dirichlet process mixtures for partially exchangeable data.
}
\keyword{internal}
