% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cDDP}
\alias{cDDP}
\title{C++ function to estimate DDP models with 1 grouping variables}
\arguments{
\item{data}{a vector of observations.}

\item{group}{group allocation of the data.}

\item{ngr}{number of groups.}

\item{grid}{vector to evaluate the density.}

\item{niter}{number of iterations.}

\item{nburn}{number of burn-in iterations.}

\item{m0}{expectation of location component.}

\item{k0}{tuning parameter of variance of location component.}

\item{a0}{parameter of scale component.}

\item{b0}{parameter of scale component.}

\item{mass}{mass of Dirichlet process.}

\item{wei}{prior weight of the specific processes.}

\item{b}{tuning parameter of weights distribution}

\item{napprox}{number of approximating values.}

\item{n_approx_unif}{number of approximating values of the importance step for the weights updating.}

\item{nupd}{number of iterations to show current updating.}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE).}

\item{print_message}{print the status.}

\item{light_dens}{if TRUE return only the posterior mean of the density}
}
\description{
C++ function to estimate DDP models with 1 grouping variables
}
\keyword{internal}
