% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSLI}
\alias{cSLI}
\title{C++ function to estimate Pitman-Yor univariate mixtures via slice sampler - LOCATION SCALE}
\arguments{
\item{data}{a vector of observations}

\item{grid}{vector to evaluate the density}

\item{niter}{number of iterations}

\item{nburn}{number of burn-in iterations}

\item{m0}{expectation of location component}

\item{k0}{tuning parameter of variance of location component}

\item{a0}{parameter of scale component}

\item{b0}{parameter of scale component}

\item{m1}{mean of hyperdistribution of m0}

\item{s21}{variance of hyperdistribution of m0}

\item{tau1}{shape parameter of hyperdistribution of k0}

\item{tau2}{rate parameter of hyperdistribution of k0}

\item{a1}{shape parameter of hyperdistribution of b0}

\item{b1}{rate parameter of hyperdistribution of b0}

\item{mass}{parameter}

\item{nupd}{number of iterations to show current updating}

\item{out_param}{if TRUE, return also the location and scale paramteres lists}

\item{out_dens}{if TRUE, return also the estimated density (default TRUE)}

\item{sigma_PY}{second parameter of PY}

\item{print_message}{print the status}

\item{hyper, }{if TRUE use hyperpriors, default TRUE}

\item{indep, }{if TRUE use the independent slice efficient}
}
\description{
C++ function to estimate Pitman-Yor univariate mixtures via slice sampler - LOCATION SCALE
}
\keyword{internal}
