% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{plot.modCond}
\alias{plot.modCond}
\title{modCond summary method}
\usage{
\method{plot}{modCond}(x, dimension = c(1, 2), col = "#0037c4", ...)
}
\arguments{
\item{x}{an object of class modCond}

\item{dimension}{if multivariate, the two dimensions for the plot (if they are equal, a marginal plot is performed)}

\item{col}{the color of the lines}

\item{...}{additional arguments to be passed}
}
\description{
modCond summary method
}
