% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{summary.BNPdens}
\alias{summary.BNPdens}
\title{BNPdens summary method}
\usage{
\method{summary}{BNPdens}(object, ...)
}
\arguments{
\item{object}{an object of class \code{BNPdens};}

\item{...}{additional arguments to be passed.}
}
\description{
BNPdens summary method
}
\examples{
data_toy <- c(rnorm(100, -3, 1), rnorm(100, 3, 1))
grid <- seq(-7, 7, length.out = 50)
est_model <- PYdensity(y = data_toy, mcmc = list(niter = 100,
                      nburn = 10, napprox = 10), output = list(grid = grid))
class(est_model)
summary(est_model)

}
