% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_generic.R
\name{summary.modCond}
\alias{summary.modCond}
\title{modCond summary method}
\usage{
\method{summary}{modCond}(object, ...)
}
\arguments{
\item{object}{an object of class modCond}

\item{...}{additional arguments to be passed}
}
\description{
modCond summary method
}
